pyproject.toml
[tool.poetry]
name = "word-find"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
ipython = "^8.31.0"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
rest-framework-simplejwt = "^0.0.2"
docker = "^7.1.0"
gunicorn = "^23.0.0"
elasticsearch = "^8.17.0"
elasticsearch-dsl = "^8.17.0"
coverage = "^7.6.10"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
